去除datacenter模块中web部分 添加kafka模块 数据库表ID扩展为Long(mysql中对应bigint)类型

+++ database/src/main/java/com/xidian/iot/database/mapper/NodeAttrStdMapper.java
@@ -10,7 +10,7 @@ public interface NodeAttrStdMapper {
 
     int deleteByExample(NodeAttrStdExample example);
 
-    int deleteByPrimaryKey(Integer nasId);
+    int deleteByPrimaryKey(Long nasId);
 
     int insert(NodeAttrStd record);
 
@@ -18,7 +18,7 @@ public interface NodeAttrStdMapper {
 
     List<NodeAttrStd> selectByExample(NodeAttrStdExample example);
 
-    NodeAttrStd selectByPrimaryKey(Integer nasId);
+    NodeAttrStd selectByPrimaryKey(Long nasId);
 
     int updateByExampleSelective(@Param("record") NodeAttrStd record, @Param("example") NodeAttrStdExample example);
