+++ database/src/main/java/com/xidian/iot/database/pojo/NodeCmdGroup.java
@@ -4,18 +4,18 @@ package com.xidian.iot.database.pojo;
  * 节点命令组，发不同类型的控制命令时，会对应不同的的处理方式，该表对控制命令进行了分类，比如“视频”“转动““频率“等不同类别的动作；一个命令对应到一个命令组，一个命令组可包含多条命令
  *
  * @author mrl
- * @date   2020/09/01
+ * @date   2020/09/09
  */
 public class NodeCmdGroup {
     /**
      * 节点命令组ID
      */
-    private Integer ncgId;
+    private Long ncgId;
 
     /**
      * 节点ID，一个命令组对应一个节点，具体关系待商榷
      */
-    private Integer nodeId;
+    private Long nodeId;
 
     /**
      * 节点命令组名称
@@ -27,19 +27,19 @@ public class NodeCmdGroup {
      */
     private String ncgKey;
 
-    public Integer getNcgId() {
+    public Long getNcgId() {
         return ncgId;
     }
 
-    public void setNcgId(Integer ncgId) {
+    public void setNcgId(Long ncgId) {
         this.ncgId = ncgId;
     }
 
-    public Integer getNodeId() {
+    public Long getNodeId() {
         return nodeId;
     }
 
-    public void setNodeId(Integer nodeId) {
+    public void setNodeId(Long nodeId) {
         this.nodeId = nodeId;
     }
