+++ database/src/main/java/com/xidian/iot/database/pojo/NodeCmdStd.java
@@ -4,13 +4,13 @@ package com.xidian.iot.database.pojo;
  * 节点命令名称模版，node_cmd_std.ncs_name字段对应node_cmd.nc_name，供添加命令选择“控制名称”
  *
  * @author mrl
- * @date   2020/09/01
+ * @date   2020/09/09
  */
 public class NodeCmdStd {
     /**
      * 节点命令模版ID
      */
-    private Integer ncsId;
+    private Long ncsId;
 
     /**
      * 命令属性
@@ -22,11 +22,11 @@ public class NodeCmdStd {
      */
     private String ncsName;
 
-    public Integer getNcsId() {
+    public Long getNcsId() {
         return ncsId;
     }
 
-    public void setNcsId(Integer ncsId) {
+    public void setNcsId(Long ncsId) {
         this.ncsId = ncsId;
     }
